Removed the rubbish state fields from FiComClient. Put them in their right place inside FiComResponse.
They were rubbish, since the same FiComClient object can be shared by multiple threads and whatnot. The state of a given transaction is not the business of the general client. It's the business of the particular transaction, represented by the final variables inside the big method call and by the message objects themselves.