Improve Expected server TLS certificate data management
Using the MssClient(MsspConf) constructor the system does pick up a set
of certificates, out of which one is expected to be seen at the server.
If there is no truststore parameter in the configuration, or the
truststore is empty, then any server certificate is accepted.
These certificates are not validated online in any way, only that server
certificate can be found in this set.